Outside scene, changed lights, added some furniture, made some textures, FIXED SCALE ISSUES
A huge issue is scale. You have to make sure that all of your things are scared to the real scale in life. The room was very short and a number of other things were very small or very large. But the room looks way more realistic now and gives you a sense of presents.